--- a/addons/board/static/tests/board.test.js
+++ b/addons/board/static/tests/board.test.js
@@ -608,21 +608,21 @@ describe("board_desktop", () => {
         onRpc(({ method, kwargs }) => {
             if (method === "get_property_definition") {
                 return {};
-            } else if (
-                method === "formatted_read_group" &&
-                kwargs.groupby?.includes("properties.my_char")
-            ) {
+            } else if (method === "formatted_read_grouping_sets") {
                 return [
-                    {
-                        "properties.my_char": false,
-                        __extra_domain: [["properties.my_char", "=", false]],
-                        __count: 2,
-                    },
-                    {
-                        "properties.my_char": "aaa",
-                        __extra_domain: [["properties.my_char", "=", "aaa"]],
-                        __count: 1,
-                    },
+                    [{ __count: 3, __extra_domain: [] }],
+                    [
+                        {
+                            "properties.my_char": false,
+                            __extra_domain: [["properties.my_char", "=", false]],
+                            __count: 2,
+                        },
+                        {
+                            "properties.my_char": "aaa",
+                            __extra_domain: [["properties.my_char", "=", "aaa"]],
+                            __count: 1,
+                        },
+                    ],
                 ];
             }
         });