--- a/addons/hr_skills/static/src/views/pivot/skills_pivot_model.js
+++ b/addons/hr_skills/static/src/views/pivot/skills_pivot_model.js
@@ -1,6 +1,5 @@
 import { PivotModel } from "@web/views/pivot/pivot_model";
 
-
 export class SkillsPivotModel extends PivotModel {
     setup() {
         super.setup(...arguments);
@@ -8,28 +7,26 @@ export class SkillsPivotModel extends PivotModel {
         this.departmentNamesById = {};
     }
 
-    async _getSubGroups(groupBy, params) {
-        if (!this.departmentsFetch){
-            const departmentNames = await this.orm.webSearchRead(
-                "hr.department",
-                [],
-                {
-                    specification: {
-                        name: {},
-                    },
-                }
-            );
+    async _getGroupsSubdivision(params, groupInfo) {
+        if (!this.departmentsFetch) {
+            const departmentNames = await this.orm.webSearchRead("hr.department", [], {
+                specification: {
+                    name: {},
+                },
+            });
             this.departmentNamesById = Object.fromEntries(
                 departmentNames.records.map(({ id, name }) => [id, name])
             );
         }
-        const data = await super._getSubGroups(...arguments);
-        if (groupBy.includes("department_id")) {
-            data.forEach((res) => {
-                res.department_id[1] = this.departmentNamesById[res.department_id[0]];
-            });
-        }
+        const multiData = await super._getGroupsSubdivision(...arguments);
 
-        return data;
+        for (const [groupBy, data] of params.groupingSets.map((g, i) => [g, multiData[i]])) {
+            if (groupBy.includes("department_id")) {
+                data.forEach((res) => {
+                    res.department_id[1] = this.departmentNamesById[res.department_id[0]];
+                });
+            }
+        }
+        return multiData;
     }
 }