--- a/addons/spreadsheet/static/tests/global_filters/global_filters_model.test.js
+++ b/addons/spreadsheet/static/tests/global_filters/global_filters_model.test.js
@@ -1347,18 +1347,13 @@ test("load data only once if filter is not active (without default value)", asyn
     const { model } = await createModelWithDataSource({
         spreadsheetData,
         mockRPC: function (route, { model, method, kwargs }) {
-            if (model === "partner" && method === "formatted_read_group") {
+            if (model === "partner" && method === "formatted_read_grouping_sets") {
                 expect.step(`${model}/${method}`);
             }
         },
     });
     await waitForDataLoaded(model);
-    expect.verifySteps([
-        "partner/formatted_read_group",
-        "partner/formatted_read_group",
-        "partner/formatted_read_group",
-        "partner/formatted_read_group",
-    ]);
+    expect.verifySteps(["partner/formatted_read_grouping_sets"]);
     expect(getCellValue(model, "A1")).toBe(131);
 });
 
@@ -1396,13 +1391,13 @@ test("load data only once if filter is active (with a default value)", async fun
     const { model } = await createModelWithDataSource({
         spreadsheetData,
         mockRPC: function (route, { model, method, kwargs }) {
-            if (model === "partner" && method === "formatted_read_group") {
+            if (model === "partner" && method === "formatted_read_grouping_sets") {
                 expect.step(`${model}/${method}`);
             }
         },
     });
     await waitForDataLoaded(model);
-    expect.verifySteps(["partner/formatted_read_group"]);
+    expect.verifySteps(["partner/formatted_read_grouping_sets"]);
     expect(getCellValue(model, "A1")).toBe("");
 });
 
@@ -1431,18 +1426,13 @@ test("don't reload data if an empty filter is added", async function () {
     const { model } = await createModelWithDataSource({
         spreadsheetData,
         mockRPC: function (route, { model, method, kwargs }) {
-            if (model === "partner" && method === "formatted_read_group") {
+            if (model === "partner" && method === "formatted_read_grouping_sets") {
                 expect.step(`${model}/${method}`);
             }
         },
     });
     await waitForDataLoaded(model);
-    expect.verifySteps([
-        "partner/formatted_read_group",
-        "partner/formatted_read_group",
-        "partner/formatted_read_group",
-        "partner/formatted_read_group",
-    ]);
+    expect.verifySteps(["partner/formatted_read_grouping_sets"]);
     expect(getCellValue(model, "A1")).toBe(131);
     addGlobalFilter(model, {
         id: "42",
@@ -1472,7 +1462,7 @@ test("don't load data if a filter is added but the data is not needed", async fu
     const { model } = await createModelWithDataSource({
         spreadsheetData,
         mockRPC: function (route, { model, method, kwargs }) {
-            if (model === "partner" && method === "formatted_read_group") {
+            if (model === "partner" && method === "formatted_read_grouping_sets") {
                 expect.step(`${model}/${method}`);
             }
         },
@@ -1494,7 +1484,7 @@ test("don't load data if a filter is added but the data is not needed", async fu
     expect(getCellValue(model, "A1")).toBe("Loading...");
     await animationFrame();
     expect(getCellValue(model, "A1")).toBe("");
-    expect.verifySteps(["partner/formatted_read_group"]);
+    expect.verifySteps(["partner/formatted_read_grouping_sets"]);
 });
 
 test("don't load data if a filter is activated but the data is not needed", async function () {
@@ -1522,7 +1512,7 @@ test("don't load data if a filter is activated but the data is not needed", asyn
     const { model } = await createModelWithDataSource({
         spreadsheetData,
         mockRPC: function (route, { model, method, kwargs }) {
-            if (model === "partner" && method === "formatted_read_group") {
+            if (model === "partner" && method === "formatted_read_grouping_sets") {
                 expect.step(`${model}/${method}`);
             }
         },
@@ -1539,7 +1529,7 @@ test("don't load data if a filter is activated but the data is not needed", asyn
     expect(getCellValue(model, "A1")).toBe("Loading...");
     await animationFrame();
     expect(getCellValue(model, "A1")).toBe("");
-    expect.verifySteps(["partner/formatted_read_group"]);
+    expect.verifySteps(["partner/formatted_read_grouping_sets"]);
 });
 
 test("Default value defines value", async function () {