--- a/addons/web/static/src/model/sample_server.js
+++ b/addons/web/static/src/model/sample_server.js
@@ -124,6 +124,8 @@ export class SampleServer {
                 return this._mockWebReadGroup(params);
             case "formatted_read_group":
                 return this._mockFormattedReadGroup(params);
+            case "formatted_read_grouping_sets":
+                return this._mockFormattedReadGroupingSets(params);
             case "read_progress_bar":
                 return this._mockReadProgressBar(params);
             case "read":
@@ -505,6 +507,23 @@ export class SampleServer {
             });
         }
         return result;
+    }
+
+    /**
+     * Mocks calls to the base method of formatted_read_grouping_sets method.
+     *
+     * @param {Object} params
+     * @param {string} params.model
+     * @param {string[][]} params.grouping_sets
+     * @param {string[]} params.aggregates
+     * @returns {Object[]} Object with keys groups and length
+     */
+    _mockFormattedReadGroupingSets(params) {
+        const res = [];
+        for (const groupBy of params.grouping_sets) {
+            res.push(this._mockFormattedReadGroup({ ...params, groupBy }));
+        }
+        return res;
     }
 
     /**