--- a/odoo/addons/test_read_group/tests/__init__.py
+++ b/odoo/addons/test_read_group/tests/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from . import test_formatted_read_group
 from . import test_private_read_group
+from . import test_read_grouping_sets
 from . import test_read_progress_bar
 from . import test_web_fill_temporal
 from . import test_web_group_expand