… addons/spreadsheet/static/tests/pivots/model/pivot_plugin.test.js
… addons/web/static/tests/views/pivot_view.test.js
… odoo/addons/test_read_group/tests/test_read_grouping_sets.py
… odoo/orm/models.py