Currently, opening a pivot view triggers multiple `formatted_read_group` requests. The number of these requests (X) depends on the number of expanded field columns and rows (X = (number of row fields + 1) * (number of column fields + 1)). For example, a Commission pivot view grouped by `plan_id` and `user_id` for rows, and `payment_date` for columns, would result in 6 `formatted_read_group` calls ((2 + 1) * (1 + 1)). 

These individual `formatted_read_group` requests are often slow and frequently lead to sequential scans on the target table/view. Since all these requests share the same arguments (except for the `groupby` parameter) and do not use `limit` or `offset`, they are ideal candidates for PostgreSQL's `GROUPING SETS` feature. This allows combining all these requests into a single query with comparable performance to a single previous `formatted_read_group` call. 

This commit introduces `_read_grouping_sets` and its web-client version, `formatted_read_grouping_sets`, and integrates it into the pivot model. This significantly reduces the number of heavy RPC calls to the server, making all pivot views (including those used in dashboards with spreadsheets) faster in term of user time and greatly reduces the
worker time taken for pivots/dashboards views.

### Performance

###### Local: In Sales > Reporting > Commission:

Test on my compute with sale_commission, purchase, mrp modules + populate subcommand and in multi workers=6

**Before - 6 RPC:**
User Time: 6.051 +- 0.033 sec
Worker time: 22.230 -+ 0.276 sec

**After - 1 RPC:**
User Time = Worker time: 1.988 +- 0.016 sec

###### Next: In Sales > Reporting > Commission:
Test on real data on our database:

**Before - 6 RPC:** 
User Time: 25.84 sec
Worker time: 79.78 sec

**After - 1 RPC:**
User Time = Worker time: 13.80 sec

###### Next: Dashboard > CRM - Leads:

**Before - 28 RPC:** 
User Time: -+ 40 sec
Worker time: 241.4 sec

**After - 14 RPC:**
User Time: -+ 25 sec
Worker time: 146.82 sec

https://github.com/odoo/enterprise/pull/89396

Special thanks to @Polymorphe57 for the idea long time ago.