@robodoo 2025-01-20
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/194413.png)](https://mergebot.odoo.com/odoo/odoo/pull/194413)